(refactor) Move draft rules to a flat src.draft_rules module
